+++ database/migrations/2023_05_24_101620_create_bill_items_table.php
@@ -16,9 +16,9 @@ return new class extends Migration
             $table->unsignedBigInteger('bill_id');
             $table->string('sku')->nullable();
             $table->text('description')->nullable();
-            $table->float('quantity')->default(0.0000);
-            $table->float('rate', 8, 4)->default(0.0000);
-            $table->float('total', 8, 4)->default(0.0000);
+            $table->float('quantity', 8, 4)->default(0.0000);
+            $table->double('rate')->default(0);
+            $table->double('total')->default(0);
             $table->integer('company_id');
             $table->integer('created_by')->nullable();
             $table->integer('updated_by')->nullable();
