+++ database/migrations/2023_05_25_123602_create_bom_items_table.php
@@ -19,8 +19,8 @@ return new class extends Migration
             $table->string('name')->nullable();
             $table->string('unit')->nullable();
             $table->float('quantity', 8, 4)->default(0.0000);
-            $table->float('rate', 8, 4)->default(0.0000);
-            $table->float('total', 8, 4)->default(0.0000);
+            $table->double('rate')->default(0);
+            $table->double('total')->default(0);
             $table->integer('company_id');
             $table->integer('created_by')->nullable();
             $table->integer('updated_by')->nullable();
